requirements.txt
# MT5-CRS Project Dependencies
# Last Updated: 2025-12-27 (Work Order #023.9 - ML Stack Upgrade)

# ============================================================================
# Core Infrastructure
# ============================================================================
requests>=2.28.0
python-dotenv>=0.19.0
curl_cffi>=0.7.0         # Python 3.7+ (restored for Python 3.9 - Task #040.13)

# NOTE: curl_cffi was commented for Python 3.6 compatibility
#       Restored in Task #040.13 after upgrading to Python 3.9
#       Graceful fallback still available: gemini_review_bridge.py lines 23-28

# ============================================================================
# Work Order #022: ZeroMQ Communication Fabric
# ============================================================================
pyzmq>=25.0.0              # High-performance messaging (ZeroMQ)
psutil>=5.9.0              # Process and system utilities

# ============================================================================
# Work Order #023.9: Modern ML Stack (Production)
# ============================================================================

# Data Processing & Analysis
pandas>=2.0.0              # DataFrame operations (2.x for performance)
numpy>=1.24.0              # Numerical computing
scipy>=1.10.0              # Scientific computing

# Machine Learning - Core
scikit-learn>=1.3.0        # ML algorithms and preprocessing
xgboost>=2.0.0             # Gradient boosting (2.x for GPU support)
lightgbm>=4.0.0            # Gradient boosting (Microsoft)

# Model Persistence & Serving
joblib>=1.3.0              # Model serialization
cloudpickle>=2.2.0         # Enhanced pickling

# Feature Engineering
statsmodels>=0.14.0        # Statistical models
# Note: Technical indicators (RSI, MACD, Bollinger) implemented using pandas/numpy
# pandas_ta requires Python 3.12+ (unavailable on Python 3.9)

# Testing & Quality
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Coverage reporting

# ============================================================================
# Work Order #032: Data Nexus Infrastructure
# ============================================================================
sqlalchemy>=2.0.23         # ORM for TimescaleDB
psycopg2-binary>=2.9.9    # PostgreSQL adapter
redis>=5.0.1               # Redis client (sync + async support via redis.asyncio)
# python-dotenv>=1.0.0     # Already listed above

# ============================================================================
# Work Order #034: Async EODHD Ingestion Engine
# ============================================================================
aiohttp>=3.9.0             # Async HTTP client
aiodns>=3.1.0              # Async DNS resolution
click>=8.1.0               # CLI argument parsing

# ============================================================================
# Work Order #036: Real-time WebSocket Engine
# ============================================================================
websockets>=15.0.0         # Async WebSocket client

# ============================================================================
# Notion Integration
# ============================================================================
notion-client>=2.0.0       # Notion API client

# ============================================================================
# Optional Dependencies (Development/Research Only)
# ============================================================================
# Install with: pip install -r requirements-dev.txt
#
# Deep Learning:
# - tensorflow>=2.13.0       # Deep learning (TensorFlow)
# - torch>=2.0.0             # Deep learning (PyTorch)
#
# Technical Analysis:
# - ta-lib>=0.4.0            # Requires system lib installation
#
# Visualization:
# - matplotlib>=3.7.0        # Plotting
# - seaborn>=0.12.0          # Statistical visualization
#
# Create requirements-dev.txt with above for development environment

# ============================================================================
# Version Lock Notes
# ============================================================================
# pandas>=2.0     : ~2x faster operations, better memory efficiency
# xgboost>=2.0    : GPU acceleration, better API
# pyzmq>=25.0     : Latest ZeroMQ features
# scikit-learn>=1.3 : Latest preprocessing pipelines
